### 2022-05-15 #4394 @arfon
@VChristiaens - thanks for your submission to JOSS. Does this paper accompany a paper in a AAS journal?

Also, please quote the citations for Carles like this:

```
  - name: Carles Cantero
    orcid: 0000-0003-2073-782X
    affiliation: "1,7"
```
