### 2022-05-16 #4394 @VChristiaens
@arfon - thanks for your prompt reply. There was a [2017 paper in AJ](https://ui.adsabs.harvard.edu/abs/2017AJ....154....7G/abstract), which presented a previous release of the package. However significant additions and changes were implemented to the package in the past 5 years, which motivate a new publication - this time in JOSS. So to answer your question: no other AAS journal paper apart from the existing 2017 one.

Thanks for the fix regarding citation.
