### 2022-05-18 f1a7b36
minor bug fix for NaN correction in 2D arrays
