+++ vip_hci/preproc/recentering.py
@@ -615,9 +615,10 @@ def cube_recenter_satspots(array, xy, subi_size=19, sigfactor=6, plot=True,
         return array_rec
 
 
-def frame_center_radon(array, cropsize=None, hsize=0.4, step=0.01,
-                       mask_center=None, nproc=None, satspots_cfg=None, 
-                       theta_0=0, delta_theta=5, imlib='vip-fft', 
+def frame_center_radon(array, cropsize=None, hsize_ini=1., step_ini=0.1, 
+                       n_iter=5, tol=0.05, mask_center=None, nproc=None, 
+                       satspots_cfg=None, theta_0=0, delta_theta=5, 
+                       gauss_fit=True, imlib='vip-fft', 
                        interpolation='lanczos4', full_output=False, 
                        verbose=True, plot=True, debug=False):
     """ Finding the center of a broadband (co-added) frame with speckles and
@@ -632,11 +633,22 @@ def frame_center_radon(array, cropsize=None, hsize=0.4, step=0.01,
         Size in pixels of the cropped central area of the input array that will
         be used. It should be large enough to contain the bright elongated
         speckle or satellite spots.
-    hsize : float, optional
-        Size of the box for the grid search. The frame is shifted to each
-        direction from the center in a hsize length with a given step.
-    step : float, optional
-        The step of the coordinates change.
+    hsize_ini : float, optional
+        Size of the box for the grid search for first centering iteration. The 
+        frame is shifted to each direction from the center in a hsize length 
+        with a given step.
+    step_ini : float, optional
+        The step of the coordinates change in the first step. Note: should not 
+        be too fine for efficiency as it is automatically refined at each step.
+    n_iter : int, optional
+        Number of iterations for finer recentering. At each step, a finer 
+        step is considered based on the amplitude of the shifts found in the 
+        previous step. Iterations are particularly relevant when mask_center is 
+        not None (as the masked area will change from one iteration to the next).
+    tol : float, optional
+        Absolute tolerance on relative shift from one iteration to the next to
+        consider convergence. If the absolute value of the shift is found to be 
+        less than tol, the iterative algorithm is stopped.
     mask_center : None or int, optional
         If None the central area of the frame is kept. If int a centered zero
         mask will be applied to the frame. By default the center isn't masked.
@@ -647,9 +659,9 @@ def frame_center_radon(array, cropsize=None, hsize=0.4, step=0.01,
         If satellite spots are present, provide a string corresponding to the
         configuration of the satellite spots: as a cross ('x'), as a
         plus sign ('+') or 'custom' (provide theta_0). Leave to None if no 
-        satellite spots present. Usually the Radon transform centering works 
-        better if bright satellite spots are present.
-    theta_0: float, optional
+        satellite spots present. Note: setting satspots_cfg to non-None value
+        leads to varying performance depending on dataset.
+    theta_0: float between [0,90[, optional
         Azimuth of the first satellite spot. Only considered if satspots_cfg is
         set to 'custom'.
     delta_theta: float, optional
@@ -662,7 +674,7 @@ def frame_center_radon(array, cropsize=None, hsize=0.4, step=0.01,
     interpolation : str, optional
         See the documentation of the ``vip_hci.preproc.frame_shift`` function.
     full_output: bool, optional
-        Whether to also return the cost map.
+        Whether to also return the cost map, and uncertainty on centering.
     verbose : bool optional
         Whether to print to stdout some messages and info.
     plot : bool, optional
@@ -672,153 +684,228 @@ def frame_center_radon(array, cropsize=None, hsize=0.4, step=0.01,
 
     Returns
     -------
-    [full_output=True] 2d np array
-        Radon cost function surface is returned if full_output set to True
     optimy, optimx : float
[…]
