### 2022-05-27 96cd31e
Removed redundant argument description in nmf_annular
