### 2022-05-27 fdf7750
Added mask post-rotation for pca_adi
