### 2022-05-28 d564d17
Added 'greedy' module
