+++ vip_hci/invprob/__init__.py
@@ -1,3 +1,3 @@
-
-
 from .andromeda import *
+from .fmmf import *
+from .paco import *
