### 2022-05-28 ecf3d9a
adapted import to stim_map and inverse_stim_map
