### 2022-05-28 3c560c4
Tests: Updated arguments of frame_center_radon according to recent update of the function
