### 2022-05-28 6c63b3b
Tests: Updated arguments of frame_center_radon according to recent update of the function
