### 2022-05-29 1c4c62f
Tests: Updated Radon test to take mean of spectral channels before applying Radon
