### 2022-05-29 478bf90
Tests: pre-centering with Radon before testing a random shift
