### 2022-05-30 3c4555f
Tests: pre-centering with Radon before testing a random shift
