### 2022-05-31 0da64f1
Improved docs for pca_annulus
