### 2022-05-31 c5e771e
Debugged Radon-transform based centering: added high-pass filtering option
