### 2022-05-31 9022484
Test: solved poor-performance Radon-transform based centering by subtracting biasing BKG star and high-pass filtering input image
