### 2022-05-31 7f7fa39
Test: solved poor-performance Radon-transform based centering by subtracting biasing BKG star and high-pass filtering input image
