### 2022-06-01 b218538
Updated version associated to publication
