### 2022-06-01 1bb43f5
tests: changed npc for NEGFC-MCMC test to a more optimal value
