### 2022-06-02 #4394 @christinahedges
Hi @VChristiaens, thanks for your submission to JOSS! Thank you for being patient while this tool was waitlisted. I'm excited to get this review kicked off, this seems like a large and well used astronomy tool. I'm going to start by reading through your submission, and then I will ping some potential reviewers. We will find at least two reviewers, and then start the review. They will go through a checklist and then provide you feedback. Here are some quick comments:

1. I see you have an existing publication for an earlier version of this code (v0.7), I am going to take some time to check in with the editorial team that there's nothing special we would like to do in this case. I'll report back.
2. One quick easy fix I have noticed in your submission paper is there isn't a discussion of similar tools in the field, and why this tool provides additional usability/functionality for the community. Would you be able to add a paragraph to your statement of need discussing this?

I will start looking for reviewers. I have some potential reviewers in mind, but sometimes it can be hard to find reviewers who have no [conflicts of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy), particularly in the case of a large tool. If you know of researchers who have relevant expertise and are not conflicted with you, please feel free to suggest their names here, otherwise I will search for reviewers and get back to you.

It can sometimes take a while to get together reviewers, but after we find them the review is usually complete in 4-6 weeks.
