### 2022-06-03 fd879c5
Updated ref in optimal iterative PCA algo description
