+++ paper.bib
@@ -1,11 +1,82 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Valentin Christiaens at 2022-05-03 23:07:04 +0200 
-
-%% Saved with string encoding Unicode (UTF-8)
+%% Created for Valentin Christiaens at 2022-06-03 17:47:14 +0200 
 
 
+%% Saved with string encoding Unicode (UTF-8) 
+
+
+
+@article{pynpoint:2015,
+	abstract = {We announce the public release of PynPoint, a Python package that we have developed for analysing exoplanet data taken with the angular differential imaging observing technique. In particular, PynPoint is designed to model the point spread function of the central star and to subtract its flux contribution to reveal nearby faint companion planets. The current version of the package does this correction by using a principal component analysis method to build a basis set for modelling the point spread function of the observations. We demonstrate the performance of the package by reanalysing publicly available data on the exoplanet β Pictoris b, which consists of close to 24,000 individual image frames. We show that PynPoint is able to analyse this typical data in roughly 1.5 min on a Mac Pro, when the number of images is reduced by co-adding in sets of 5. The main computational work, the calculation of the Singular-Value-Decomposition, parallelises well as a result of a reliance on the SciPy and NumPy packages. For this calculation the peak memory load is 6 GB, which can be run comfortably on most workstations. A simpler calculation, by co-adding over 50, takes 3 s with a peak memory usage of 600 MB. This can be performed easily on a laptop. In developing the package we have modularised the code so that we will be able to extend functionality in future releases, through the inclusion of more modules, without it affecting the users application programming interface. We distribute the PynPoint package under GPLv3 licence through the central PyPI server, and the documentation is available online (http://pynpoint.ethz.ch).},
+	adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+	adsurl = {https://ui.adsabs.harvard.edu/abs/2015A&C....10..107A},
+	archiveprefix = {arXiv},
+	author = {{Amara}, A. and {Quanz}, S.~P. and {Akeret}, J.},
+	date-added = {2022-06-03 17:47:00 +0200},
+	date-modified = {2022-06-03 17:47:14 +0200},
+	doi = {10.1016/j.ascom.2015.01.003},
+	eprint = {1405.3284},
+	journal = {Astronomy and Computing},
+	keywords = {Methods: data analysis, Techniques: image processing, Planets and satellites: detection, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+	month = apr,
+	pages = {107-115},
[…]
