### 2022-06-03 #4394 @christinahedges
@VChristiaens Thanks again for your submission. Before we proceed, I'd like to ask for you to post here a detailed description of the changes made since v0.7 of VIP that are now implemented in this version. This will allow the editorial team to verify this is a substantial new contribution over your old publication, so that it meets the submission criteria of JOSS. So long as from your list we are able to establish it is a substantial new contribution, we'll be able to proceed with your submission. After we've done this check, I'll get some reviewers together.
