### 2022-06-07 4a86da9
Addition of the iterative spectral deconvolution algorithm for bad pixel correction
