+++ vip_hci/preproc/badpixremoval.py
@@ -2,6 +2,14 @@
 
 """
 Module with functions for correcting bad pixels in cubes.
+
+.. [AAC01]
+   | Aach & Metzler 2001
+   | **Defect interpolation in digital radiography how
+object-oriented transform coding helps**
+   | *SPIE, Proceedings Volume 4322, Medical Imaging 2001: Image Processing*
+   | `https://doi.org/10.1117/12.431161
+     <https://doi.org/10.1117/12.431161>`_
 """
 
 
@@ -11,17 +19,20 @@ __all__ = ['frame_fix_badpix_isolated',
            'cube_fix_badpix_annuli',
            'cube_fix_badpix_clump',
            'cube_fix_badpix_ifs',
-           'cube_fix_badpix_with_kernel']
+           'cube_fix_badpix_interp',
+           'frame_fix_badpix_fft']
 
 import numpy as np
 from skimage.draw import disk, ellipse
 from scipy.ndimage import median_filter
 from astropy.stats import sigma_clipped_stats
-from ..stats import sigma_filter
+from multiprocessing import cpu_count
+from ..stats import clip_array, sigma_filter
 from ..var import frame_center, get_annulus_segments, frame_filter_lowpass
-from ..stats import clip_array
 from ..config import timing, time_ini, Progressbar
+from ..config.utils_conf import pool_map, iterable
 from .rescaling import find_scal_vector, frame_rescaling
+from .cosmetics import frame_pad
 
 import warnings
 try:
@@ -127,7 +138,7 @@ def frame_fix_badpix_isolated(array, bpm_mask=None, sigma_clip=3, num_neig=5,
     count_bp = np.sum(bpm_mask)
 
     if verbose:
-        msg = "/nDone replacing {} bad pixels using the median of neighbors"
+        msg = "Done replacing {} bad pixels using the median of neighbors"
         print(msg.format(count_bp))
         timing(start)
 
@@ -1022,12 +1033,13 @@ def cube_fix_badpix_ifs(array, lbdas, fluxes=None, mask=None, cy=None, cx=None,
         return array_out
 
 
-def cube_fix_badpix_with_kernel(array, bpm_mask, mode='gauss', fwhm=4.,
-                                kernel_sz=None, psf=None, half_res_y=False,
-                                **kwargs):
+def cube_fix_badpix_interp(array, bpm_mask, mode='fft', fwhm=4., kernel_sz=None, 
+                           psf=None, half_res_y=False, nit=500, tol=1, nproc=1, 
+                           **kwargs):
     """
-    Function to correct clumps of bad pixels by interpolation with a
-    user-defined kernel (through astropy.convolution). A bad pixel map must be
+    Function to correct clumps of bad pixels by interpolation with either a
+    user-defined kernel (through astropy.convolution) or through the FFT-based
+    algorithm described in [AAC01]. A bad pixel map must be
     provided (e.g. found with function `cube_fix_badpix_clump`).
 
 
@@ -1039,7 +1051,7 @@ def cube_fix_badpix_with_kernel(array, bpm_mask, mode='gauss', fwhm=4.,
         Input bad pixel array. Should have same x,y dimenstions as array.
         If 2D, but input array is 3D, the same bpix_map will be assumed for all
         frames.
-    mode: str, optional {'gauss', 'psf'}
+    mode: str, optional {'fft', 'gauss', 'psf'}
         Can be either a 2D Gaussian ('gauss') or an input normalized PSF
         ('psf').
     fwhm: float or 1D array, opt
@@ -1060,6 +1072,15 @@ def cube_fix_badpix_with_kernel(array, bpm_mask, mode='gauss', fwhm=4.,
         there are always 2 rows of pixels with exactly the same values.
         If so, the Gaussian kernel will also be squashed vertically by a
[…]
