### 2022-06-07 85e8dd3
Tests: added test for new iterative spectral deconvolution algorithm for bad pixel correction
