### 2022-06-07 c6652b5
Paper: added mention to the new iterative spectral deconvolution bad pixel correction algorithm
