### 2022-06-08 cb1de88
Added all references of papers used in VIP, in relevant functions' documentation
