+++ paper.bib
@@ -1,12 +1,30 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Valentin Christiaens at 2022-06-07 18:16:14 +0200 
+%% Created for Valentin Christiaens at 2022-06-08 18:05:35 +0200 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
 
 
+
+@misc{pyklip:2015,
+	abstract = {pyKLIP subtracts out the stellar PSF to search for directly-imaged exoplanets and disks using a Python implementation of the Karhunen-Lo{\`e}ve Image Projection (KLIP) algorithm. pyKLIP supports ADI, SDI, and ADI+SDI to model the stellar PSF and offers a large array of PSF subtraction parameters to optimize the reduction. pyKLIP relies on a minimal amount of dependencies (numpy, scipy, and astropy) and parallelizes the KLIP algorithm to speed up the reduction. pyKLIP supports GPI and P1640 data and can interface with other data sources with the addition of new modules. It also can inject simulated planets and disks as well as automatically search for point sources in PSF-subtracted data.},
+	adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+	adsurl = {https://ui.adsabs.harvard.edu/abs/2015ascl.soft06001W},
+	archiveprefix = {ascl},
+	author = {{Wang}, Jason J. and {Ruffio}, Jean-Baptise and {De Rosa}, Robert J. and {Aguilar}, Jonathan and {Wolff}, Schuyler G. and {Pueyo}, Laurent},
+	date-added = {2022-06-08 18:04:30 +0200},
+	date-modified = {2022-06-08 18:04:37 +0200},
+	eid = {ascl:1506.001},
+	eprint = {1506.001},
+	howpublished = {Astrophysics Source Code Library, record ascl:1506.001},
+	keywords = {Software},
+	month = jun,
+	pages = {ascl:1506.001},
+	title = {{pyKLIP: PSF Subtraction for Exoplanets and Disks}},
+	year = 2015,
+	bdsk-url-1 = {https://ui.adsabs.harvard.edu/abs/2015ascl.soft06001W}}
 
 @inproceedings{Aach:2001,
 	abstract = {Today's solid state flat panel radiography detectors provide images which contain artifacts caused by lines, columns and clusters of inactive pixels. If not too large, such defects can be filled by interpolation algorithms which usually work in the spatial domain. This paper describes an alternative spectral domain approach to defect interpolation. The acquired radiograph is modeled as the undistorted image multiplied by a known binary defect window. The window effect is then removed by deconvolving the window spectrum from the spectrum of the observed, distorted radiograph. The basic ingredient of our interpolation algorithm is an earlier approach to block transform coding of arbitrarily shaped image segments, that extrapolates the segment internal intensities over a block into which the segment is embedded. For defect interpolation, the arbitrarily shaped segment is formed by a local image region with defects, thus turning extrapolation into defect interpolation. Our algorithm reconstructs both oriented structures and noise- like information in a natural-looking manner, even for large defects. Moreover, our concept can also be applied to non- binary defect windows, e.g. for gain correction.},
@@ -74,14 +92,14 @@
 	bdsk-url-2 = {https://ui.adsabs.harvard.edu/abs/2019A&A...621A..59S},
 	bdsk-url-3 = {https://ui.adsabs.harvard.edu/link_gateway/2019A&A...621A..59S/EPRINT_HTML}}
 
-@article{Gomez2016,
+@article{Gomez:2016,
[…]
