### 2022-06-08 #4394 @VChristiaens
Hi @christinahedges thank you for your feedback! I have now added a discussion of similar tools in the first paragraph of the statement of need, as requested. I have also listed a detailed list of all changes brought to the VIP package since v0.7 in [this announcement for all VIP users](https://github.com/vortex-exoplanet/VIP/discussions/512) (it is very long, but please let me know if you absolutely need it included to this thread and I can copy-paste it). Note that this announcement was intended for the VIP community before receiving your feedback - I have now added 2 more comments to the main announcement to be complete (I figured they could also be useful for the VIP community). The suggested way for you to read is:
1.  to start with the first comment (changes between v0.7 and v1.0.0);
2. then with the main announcement (changes since v1.0.0, up to the version at the moment of the announcement, i.e. v1.1.3);
3. finish with the second comment (changes between v1.1.3 and v1.3.1 - i.e. the current version, hopefully to be associated with the JOSS publication).

I've tried to summarize all changes at best, but since there have been >1000 commits by a number of different contributors  since v0.7, it is difficult to do better than that. Hopefully, this corresponds to what you need. Best wishes.
