### 2022-06-08 #4394 @VChristiaens
Since the definition used by JOSS for potential conflict of interest is quite strict, it is difficult for me to come up with a few people that haven't collaborated with any of the VIP co-authors in the past 4 years... Checking among co-authors of the two most similar tools to VIP ([``pynpoint``](https://ui.adsabs.harvard.edu/abs/2019A%26A...621A..59S/abstract) and [``pyklip``](https://ui.adsabs.harvard.edu/abs/2015ascl.soft06001W/abstract)), I think A. Amara may be one of very few options for whom there wouldn't be a conflict of interest.
