### 2022-06-09 5e11b01
updated normalize_psf and 2d fit functions to catch failed 2D fit, and in the case of 3D input PSFs (IFS), replace FWHM with interpolation of good neihbour values
