+++ vip_hci/var/fit_2d.py
@@ -26,6 +26,7 @@ from hciplot import plot_frames
 from astropy.modeling import models, fitting
 from astropy.stats import (gaussian_sigma_to_fwhm, gaussian_fwhm_to_sigma,
                            sigma_clipped_stats)
+import warnings
 from .coords import frame_center
 from .shapes import get_square
 from ..config import check_array
@@ -246,13 +247,27 @@ def fit_2dgaussian(array, crop=False, cent=None, cropsize=15, fwhmx=4, fwhmy=4,
 
     # compute uncertainties
     if fitter.fit_info['param_cov'] is not None:
-        perr = np.sqrt(np.diag(fitter.fit_info['param_cov']))
-        amplitude_e, mean_x_e, mean_y_e, fwhm_x_e, fwhm_y_e, theta_e = perr
-        fwhm_x_e /= gaussian_fwhm_to_sigma
-        fwhm_y_e /= gaussian_fwhm_to_sigma
+        with np.errstate(invalid='raise'):
+            try:
+                perr = np.sqrt(np.diag(fitter.fit_info['param_cov']))
+                amplitude_e, mean_x_e, mean_y_e, fwhm_x_e, fwhm_y_e, theta_e = perr
+                fwhm_x_e /= gaussian_fwhm_to_sigma
+                fwhm_y_e /= gaussian_fwhm_to_sigma
+            except:
+                # this means the fit failed
+                mean_y, mean_x = np.nan, np.nan
+                fwhm_y, fwhm_x = np.nan, np.nan
+                amplitude, theta = np.nan, np.nan
+                mean_y_e, mean_x_e = np.nan, np.nan
+                fwhm_y_e, fwhm_x_e = np.nan, np.nan
+                amplitude_e, theta_e = np.nan, np.nan
     else:
-        amplitude_e, theta_e, mean_x_e = None, None, None
-        mean_y_e, fwhm_x_e, fwhm_y_e = None, None, None
+        # this also means the fit failed
+        mean_y, mean_x = np.nan, np.nan
+        fwhm_y, fwhm_x = np.nan, np.nan
+        amplitude, theta = np.nan, np.nan
+        amplitude_e, theta_e, mean_x_e = np.nan, np.nan, np.nan
+        mean_y_e, fwhm_x_e, fwhm_y_e = np.nan, np.nan, np.nan
 
     if debug:
         if threshold:
@@ -395,21 +410,28 @@ def fit_2dmoffat(array, crop=False, cent=None, cropsize=15, fwhm=4,
 
     # compute uncertainties
     if fitter.fit_info['param_cov'] is not None:
-        perr = np.sqrt(np.diag(fitter.fit_info['param_cov']))
-        amplitude_err, mean_x_err, mean_y_err, gamma_err, alpha_err = perr
-        fwhm_err = 2*gamma_err
+        with np.errstate(invalid='raise'):
+            try:
+                perr = np.sqrt(np.diag(fitter.fit_info['param_cov']))
+                amplitude_e, mean_x_e, mean_y_e, gamma_e, alpha_e = perr
+                fwhm_e = 2*gamma_e
+            except:
+                # this means the fit failed
+                mean_y, mean_x, fwhm = np.nan, np.nan, np.nan
+                amplitude, alpha, gamma = np.nan, np.nan, np.nan
+                mean_y_e, mean_x_e, fwhm_e = np.nan, np.nan, np.nan
+                amplitude_e, alpha_e, gamma_e = np.nan, np.nan, np.nan
     else:
-        amplitude_err, mean_x_err, mean_y_err = None, None, None
-        gamma_err, alpha_err, fwhm_err = None, None, None
[…]
