### 2022-06-09 c1c3bf2
Updated case where fit 2D is to be considered failed
