### 2022-06-10 762a7b5
Formatting of documentation for readthedocs and update of references
