### 2022-06-10 2a1774a
Test: updated normalize psf test to avoid a failed 2D Gaussian fit
