### 2022-06-10 3ab4e1c
Docs: Updated documentation of filters method in Dataset and Frame classes
