### 2022-06-10 113687b
Minor bug fix in cube_planet_free for the case of a 4D input array
