### 2022-06-11 e3c5c64
Tests: minor bug fix in random noise injected to the synthetic psf image before testing normalize_psf
