### 2022-06-11 cabd654
Minor improvements to NEGFC algorithms, including full_output option for firstguess_from_coord function
