+++ vip_hci/fm/negfc_simplex.py
@@ -18,7 +18,8 @@ from ..config.utils_conf import sep
 
 
 __author__ = 'O. Wertz, C. A. Gomez Gonzalez, V. Christiaens'
-__all__ = ['firstguess']
+__all__ = ['firstguess',
+           'firstguess_from_coord']
 
 
 def firstguess_from_coord(planet, center, cube, angs, psfn, fwhm, annulus_width,
@@ -28,7 +29,8 @@ def firstguess_from_coord(planet, center, cube, angs, psfn, fwhm, annulus_width,
                           collapse='median', algo=pca_annulus, delta_rot=1,
                           algo_options={}, f_range=None, transmission=None,
                           mu_sigma=(0, 1), weights=None, plot=False,
-                          verbose=True, save=False, debug=False):
+                          verbose=True, save=False, debug=False, 
+                          full_output=False):
     """ Determine a first guess for the flux of a companion at a given position
     in the cube by doing a simple grid search evaluating the reduced chi2.
 
@@ -92,7 +94,7 @@ def firstguess_from_coord(planet, center, cube, angs, psfn, fwhm, annulus_width,
         dict (the latter are also kept as function arguments for compatibility
         with older versions of vip).
     f_range: numpy.array, optional
-        The range of flux tested values. If None, 20 values between 0 and 5000
+        The range of tested flux values. If None, 20 values between 0 and 5000
         are tested.
     transmission: numpy array, optional
         Array with 2 columns. First column is the radial separation in pixels.
@@ -114,12 +116,20 @@ def firstguess_from_coord(planet, center, cube, angs, psfn, fwhm, annulus_width,
         If True, display intermediate info in the shell.
     save: boolean, optional
         If True, the figure chi2 vs. flux is saved as .pdf if plot is also True
+    debug: bool, optional
+        Whether to print details of the grid search
+    full_output : bool, optional
+        Whether to also return the range of fluxes tested and their chi2r 
+        values.
 
     Returns
     -------
     res : tuple
         The polar coordinates and the flux(es) of the companion.
-
+    f_range: 1d numpy.array
+        [full_output=True] The range of tested flux values. 
+    chi2r: 1d numpy.array
+        [full_output=True] The chi2r values corresponding to tested flux values. 
     """
     def _grid_search_f(r0, theta0, ch, cube, angs, psfn, fwhm, annulus_width,
                        aperture_radius, ncomp, cube_ref=None, svd_mode='lapack',
@@ -202,6 +212,7 @@ def firstguess_from_coord(planet, center, cube, angs, psfn, fwhm, annulus_width,
 
     else:
         f0 = []
+        chi2r = []
         if plot:
             plt.figure(figsize=(8, 4))
             plt.title('$\\chi^2_{r}$ vs flux')
@@ -212,7 +223,7 @@ def firstguess_from_coord(planet, center, cube, angs, psfn, fwhm, annulus_width,
         for i in range(cube.shape[0]):
             if verbose:
                 print('Processing spectral channel {}...'.format(i))
-            chi2r = _grid_search_f(r0, theta0, i, cube, angs, psfn,
+            chi2r_tmp = _grid_search_f(r0, theta0, i, cube, angs, psfn,
                                    fwhm, annulus_width, aperture_radius,
                                    ncomp, cube_ref=cube_ref, svd_mode=svd_mode,
                                    scaling=scaling, fmerit=fmerit,
@@ -223,31 +234,32 @@ def firstguess_from_coord(planet, center, cube, angs, psfn, fwhm, annulus_width,
[…]
