+++ vip_hci/fm/negfc_mcmc.py
@@ -764,10 +764,10 @@ def mcmc_negfc_sampling(cube, angs, psfn, initial_state, algo=pca_annulus,
             dr = min(annulus_width/2, aperture_radius*fwhm/2)
             dth = 360./(2*np.pi*initial_state[0]/(aperture_radius*fwhm/2))
             bounds = [(initial_state[0] - dr, initial_state[0] + dr),  # radius
-                      (initial_state[1] - dth, initial_state[1] + dth)]  # angle
+                      (initial_state[1] - dth, initial_state[1] + dth)] # angle
             d0 = 2
         for i in range(dim-d0):
-            bounds.append((0, 10 * initial_state[d0+i]))  # flux
+            bounds.append((0, 5 * initial_state[d0+i]))  # flux
     # size of ball of parameters for MCMC initialization
     scal = abs(bounds[0][0]-initial_state[0])/initial_state[0]
     for i in range(dim):
