### 2022-06-14 ae46cd8
Added the option of multiprocessing in get_values_optimize (can be useful for faster firstguess)
