+++ vip_hci/greedy/pca_local.py
@@ -7,7 +7,8 @@ fashion) model PSF subtraction for ADI, ADI+SDI (IFS) and ADI+RDI datasets.
 
 __author__ = 'Valentin Christiaens'
 __all__ = ['pca_annular_it',
-           'feves']
+           'feves',
+           'feves_auto']
 
 from multiprocessing import cpu_count
 import numpy as np
@@ -942,7 +943,386 @@ def feves(cube, angle_list, cube_ref=None, ncomp=1, algo=pca_annular, n_it=2,
                 master_res_cube_, master_it_cube_nd, master_stim_cube, 
                 master_stim_ori_cube, master_stim_inv_cube, master_cube)
     else:
-        return master_frame    
+        return master_frame  
+    
+
+def feves_auto(cube, angle_list, cube_ref=None, ncomp=1, algo=pca_annular, 
+               n_it='auto', blur=True, fwhm=4, buff=1, thr=1, thr_per_ann=False, 
+               n_frac=6, asizes=None, n_segments=None,  n_neigh=0, strategy='ADI', 
+               n_br=6, radius_int=0, delta_rot=(0.1, 1),
+          svd_mode='lapack', init_svd='nndsvda', nproc=1, min_frames_lib=2, 
+          max_frames_lib=200, tol=1e-1, scaling=None, imlib='opencv', 
+          interpolation='lanczos4', collapse='median', full_output=False, 
+          verbose=True, weights=None, interp_order=2, rtol=1e-2, 
+          atol=1,  **rot_options):
+    """
+    Fractionation for Embedded Very young Exoplanet Search algorithm: Iterative
+    PCA or NMF applied in progressively more fractionated image sections.
+
+    The algorithm finds significant disc or planet signal in the final PCA
+    image, then subtracts it from the input cube (after rotation) just before
+    (and only for) projection onto the principal components. This is repeated
+    n_it times, which progressively reduces geometric biases in the image
+    (e.g. negative side lobes for ADI, radial negative signatures for SDI).
+    This is similar to the algorithm presented in Pairet et al. (2020).
+
+    The same parameters as pca_annular() can be provided. There are two
+    additional parameters related to the iterative algorithm: the number of
+    iterations (n_it) and the threshold (thr) used for the identification of
+    significant signals.
+
+    Parameters
+    ----------
+    cube : numpy ndarray, 3d or 4d
+        Input cube.
+    angle_list : numpy ndarray, 1d
+        Corresponding parallactic angle for each frame.
+    cube_ref : numpy ndarray, 3d, optional
+        Reference library cube. For Reference Star Differential Imaging.
+    ncomp : int or list of int or list of lists of int, optional
+        How many PCs are used as a lower-dimensional subspace to project the
+        target (sectors of) frames. Depends on the dimensionality of `cube`.
+        If int: same ncomp used at all iterations.
+        If a list: should have either the length of asizes (if the 
+        latter is provided) or a length equal to n_frac (otherwise).
+            - if its elements are int: ncomp used for each fractionation level
+            - if its elements are lists: additional iterations with different
+            ncomp will be done for each fractionation level.
+    algo: function, opt, {vip_hci.pca.pca_annular, vip_hci.nmf.nmf_annular}
+        Either PCA or NMF in concentric annuli, used iteratively.
+    n_it: int, opt
+        Number of iterations at each fractionation level.
+    buff: float, opt
+        Radial buffer expressed in fwhm, used to smooth final image. The feves 
+        algorithm will be repeated int(buff*fwhm), with radial increments of 
+        one pixel for the inner mask size. The final results are then the 
+        median of each of the int(buff*fwhm) results.
[…]
