### 2022-06-15 667137a
Made residual speckle noise uncertainty compatible with a 4D input cube
