+++ vip_hci/fm/negfc_speckle_noise.py
@@ -52,13 +52,15 @@ def speckle_noise_uncertainty(cube, p_true, angle_range, derot_angles, algo,
 
     Parameters
     ----------
-    cube: numpy array
-        The original ADI cube.
-    p_true: tuple or numpy array with 3 elements
+    cube: 3d or 4d numpy array
+        The original ADI or ADI+IFS cube.
+    p_true: tuple or numpy array with 3 (or more) elements
         The radial separation, position angle (from x=0 axis) and flux
         associated to a given companion candidate for which the speckle
-        uncertainty is to be evaluated. The planet will first
-        be subtracted from the cube, then used for test injections.
+        uncertainty is to be evaluated. The planet will first be subtracted 
+        from the cube, then used for test injections. For a 4D input cube, the 
+        length of ``p_true`` should be equal to 2 (for r, theta) + the number 
+        of spectral channels (flux at each wavelength).
     angle_range: 1d numpy array
         Range of angles (counted from x=0 axis, counter-clockwise) at which the
         fake companions will be injected, in [0,360[.
@@ -159,7 +161,17 @@ def speckle_noise_uncertainty(cube, p_true, angle_range, derot_angles, algo,
         print('#######################################################')
         print('')
 
-    r_true, theta_true, f_true = p_true
+    if len(p_true) == 3:
+        r_true, theta_true, f_true = p_true
+        nch = 1 
+    elif len(p_true) > 3 and cube.ndim == 4 and cube.shape[0] == len(p_true)-2:
+        r_true = p_true[0]
+        theta_true = p_true[1]
+        f_true = np.array(p_true[2:])
+        nch = cube.shape[0]
+    else:
+        msg = "cube ndim ({}) and parameter length ({}) combination not accepted"
+        raise TypeError(msg.format(cube.ndim, len(p_true)))
 
     if indep_ap:
         angle_span = angle_range[-1]-angle_range[0]
@@ -180,7 +192,13 @@ def speckle_noise_uncertainty(cube, p_true, angle_range, derot_angles, algo,
     interpolation = algo_options.get('interpolation', 'lanczos4')
 
     # FIRST SUBTRACT THE TRUE COMPANION CANDIDATE
-    planet_parameter = np.array([[r_true, theta_true, f_true]])
+    if len(p_true) == 3:
+        planet_parameter = np.array([[r_true, theta_true, f_true]])
+    else:
+        planet_parameter = np.zeros([1,3,nch])
+        planet_parameter[0,0,:] = r_true
+        planet_parameter[0,1,:] = theta_true
+        planet_parameter[0,2] = f_true
     cube_pf = cube_planet_free(planet_parameter, cube, derot_angles, psfn,
                                imlib=imlib, interpolation=interpolation,
                                transmission=transmission)
@@ -223,17 +241,21 @@ def speckle_noise_uncertainty(cube, p_true, angle_range, derot_angles, algo,
         residuals = np.concatenate((residuals, residuals2))
 
     if verbose:
-        print("residuals (offsets): ", residuals[:, 3], residuals[:, 4],
-              residuals[:, 5])
+        print("residuals (offsets): ", residuals[:, nch+2], residuals[:, nch+3],
+              residuals[:, nch+4])
 
-    p_simplex = np.transpose(np.vstack((residuals[:, 0], residuals[:, 1],
-                                        residuals[:, 2])))
-    offset = np.transpose(np.vstack((residuals[:, 3], residuals[:, 4],
-                                     residuals[:, 5])))
+    p_simp_stack = [residuals[:, 0], residuals[:, 1]]
[…]
