### 2022-06-21 a073334
Minor bug fix in PCA patch function
