+++ vip_hci/psfsub/pca_local.py
@@ -591,6 +591,8 @@ def do_pca_patch(matrix, frame, angle_list, fwhm, pa_threshold, ann_center,
 
         if data_ref.shape[0] < min_frames_lib and matrix_ref is None:
             raise RuntimeError(msg.format(len(indices_left), min_frames_lib))
+    else:
+        data_ref = None
 
     if matrix_ref is not None:
         #data_ref = None
