### 2022-07-12 ad0ff4e
Minor bug fix in non frame-by-frame mode for cube bad pixel correction
