+++ vip_hci/greedy/pca_local.py
@@ -3,6 +3,22 @@
 """
 Module with local/smart PCA (annulus or patch-wise in a multi-processing
 fashion) model PSF subtraction for ADI, ADI+SDI (IFS) and ADI+RDI datasets.
+
+.. [PAI18]
+   | Pairet et al. 2018
+   | **Reference-less algorithm for circumstellar disks imaging**
+   | *arXiv e-print:1812.01333*
+   | `https://arxiv.org/abs/1812.01333
+     <https://arxiv.org/abs/1812.01333>`_
+     
+     
+.. [PAI21]
+   | Pairet et al. 2021
+   | **MAYONNAISE: a morphological components analysis pipeline for circumstellar discs and exoplanets imaging in the near-infrared**
+   | *MNRAS, Volume 503, Issue 3, pp. 3724-3742*
+   | `https://arxiv.org/abs/2008.05170
+     <https://arxiv.org/abs/2008.05170>`_
+   
 """
 
 __author__ = 'Valentin Christiaens'
@@ -15,19 +31,21 @@ import numpy as np
 from scipy.interpolate import InterpolatedUnivariateSpline
 from ..config.utils_conf import pool_map, iterable
 from .pca_fullfr import _interp2d_rad
-from ..preproc import cube_derotate, cube_shift, cube_collapse
+from ..preproc import cube_derotate, cube_shift
 from ..preproc.derotation import _define_annuli
 from ..psfsub import nmf_annular, pca_annular
 from ..metrics import throughput, stim_map, mask_sources, inverse_stim_map
 from ..var import (get_annulus_segments, matrix_scaling, mask_circle,
-                   frame_filter_lowpass, frame_center)
+                   frame_filter_lowpass, frame_center, dist_matrix)
+# DELETE
+from ..fits import write_fits
 
 def pca_annular_it(cube, angle_list, cube_ref=None, ncomp=1, n_it=10, thr=1., 
                    thr_per_ann=True, thru_corr=False, n_neigh=0, strategy='ADI', 
                    psfn=None, n_br=6, radius_int=0, fwhm=4, asize=4, 
                    n_segments=1, delta_rot=(0.1, 1), svd_mode='lapack', nproc=1, 
                    min_frames_lib=2, max_frames_lib=200, tol=1e-1, scaling=None, 
-                   imlib='opencv', interpolation='lanczos4', collapse='median', 
+                   imlib='vip-fft', interpolation='lanczos4', collapse='median', 
                    full_output=False, verbose=True, weights=None,
                    interp_order=2, rtol=1e-2, atol=1, **rot_options):
     """
@@ -558,7 +576,7 @@ def feves(cube, angle_list, cube_ref=None, ncomp=1, algo=pca_annular, n_it=2,
           n_segments=None, thru_corr=False, n_neigh=0, strategy='ADI', 
           psfn=None, n_br=6, radius_int=0, delta_rot=(0.1, 1),
           svd_mode='lapack', init_svd='nndsvda', nproc=1, min_frames_lib=2, 
-          max_frames_lib=200, tol=1e-1, scaling=None, imlib='opencv', 
+          max_frames_lib=200, tol=1e-1, scaling=None, imlib='vip-fft', 
           interpolation='lanczos4', collapse='median', full_output=False, 
           verbose=True, weights=None, interp_order=2, rtol=1e-2, 
           atol=1, smooth=False, **rot_options):
@@ -771,22 +789,15 @@ def feves(cube, angle_list, cube_ref=None, ncomp=1, algo=pca_annular, n_it=2,
         [full_output=True] 3D array with final image from each iteration.
     pcs : numpy ndarray
         [full_output=True] Principal components from the last iteration
-    residuals_cube : numpy ndarray
[…]
