### 2022-07-12 4f0cded
Updated open_fits behaviour for headers in multiple HDUs
