### 2022-07-19 343d547
Clearer description and comment in FFT-based rotation code
