### 2022-07-26 #4394 @arfon
@RemiSoummer @anand0xff @tomasstolker @jlustigy @blacy – would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under consideration is [_VIP: A Python package for high-contrast imaging_](https://github.com/vortex-exoplanet/VIP)
