### 2022-07-26 2744b10
Minor bug fix in residual speckle noise estimation function in 4D case
