### 2022-07-27 7ad87df
Adapted residual speckle noise estimation function for compatibility with the case of forced r and PA
