### 2022-08-02 f7ad11b
negfc residual speckle noise estimation algorithm made compatible with forced position + 4D mSDI+ADI input cube
