### 2022-08-02 250ab34
NEGFC figure of merit can now be computed within an area corresponding to the intersection between circle (centered on planet position) and annulus (instead of just the circle)
