### 2022-08-05 8bcafe0
improved description of approx_stellar_position
