### 2022-08-05 d8f815c
Contrast curve: improved efficiency for algorithms that do not involve pre-scaling (single instead of double estimation of radial noise profile)
