+++ vip_hci/greedy/opt_itpca.py
@@ -1,1373 +1,0 @@
-#! /usr/bin/env python
-
-"""
-FEVES algorithm for iterative PCA/NMF in progressively fractionated images.
-"""
-
-__author__ = 'Valentin Christiaens'
-__all__ = ['feves_opt',
-           'greeds_opt']
-
-try:
-    # before skimage version '0.18' the function is skimage.measure.compare_ssim
-    from skimage.measure import compare_ssim as ssim
-except:
-    # for skimage version '0.18' or above the function is skimage.metrics.structural_similarity
-    from skimage.metrics import structural_similarity as ssim
-from multiprocessing import cpu_count
-import numpy as np
-from scipy.ndimage import gaussian_filter
-import os
-from os.path import isdir, isfile
-from .pca_local import feves
-from .pca_fullfr import pca_it
-from ..config.utils_conf import pool_map, iterable
-from ..fits import write_fits, open_fits
-from ..preproc import cube_crop_frames, frame_crop
-from ..preproc.derotation import _define_annuli
-from ..psfsub import nmf_annular, nmf, pca_annular, pca
-from ..var import frame_center, get_annulus_segments
-
-
-def feves_opt(cube, angle_list, cube_ref=None, fwhm=4, strategy='ADI', 
-              algo=pca_annular, fm='med_stim', gt_image=None, nfrac_max=6, 
-              ncomps=[1,2,3,4,5], nits=[1,2], thrs=[0,0.5], drots=[0,0.5,1], 
-              buffer=1, fiducial_results=None, thru_corr=False, n_br=6, 
-              psfn=None, starphot=1., plsc=1., svd_mode='lapack', 
-              init_svd='nndsvd', scaling=None, mask_center_px=None, 
-              imlib='opencv', interpolation='lanczos4', collapse='median', 
-              check_memory=True, nproc=1, full_output=False, verbose=True, 
-              weights=None, debug=False, path='', overwrite=True,  
-              auto_crop=False, smooth=False, **rot_options):
-    """
-    Optimal version of Fractionation for Embedded Very young Exoplanet Search 
-    algorithm: Iterative PCA or NMF applied in progressively more fractionated 
-    image sections. 
-    
-    This method is relevant for stellar residuals varying temporally and to 
-    some extent spatially.
-    
-    See complete description of the algorithm in Christiaens et al. (2022b):
-    *** INSERT LINK ***    
-
-    Note: The feves_opt routine can only be used in ADI or RDI modes.
-
-    Parameters
-    ----------
-    cube : str or numpy ndarray, 3d or 4d
-        Input cube (ADI or ADI+mSDI). If a string is given, it must correspond
-        to the path to the fits file to be opened in memmap mode (for PCA
-        incremental of ADI 3d cubes). Note: the algorithm is the most efficient
-        if the provided cube is already cropped to ~35 FWHM across. This is 
-        automatically done if auto_crop is set to True.
-    angle_list : numpy ndarray, 1d
-        Corresponding parallactic angle for each frame.
-    cube_ref : numpy ndarray, 3d, optional
-        Reference library cube. For Reference Star Differential Imaging.
-    fwhm : float, optional
-        Known size of the FHWM in pixels to be used. Default value is 4.
-    strategy: str, opt
-        Whether to do iterative ADI only ('ADI'), iterative RDI only ('RDI'), 
-        or iterative RDI and iterative ADI consecutivey ('RADI'). A reference 
-        cube needs to be provided for 'RDI' and 'RADI'.
-    algo: function, opt, {vip_hci.pca.pca, vip_hci.nmf.nmf}
-        Either PCA or NMF in concentric annuli, used iteratively.
[…]
