+++ setup.py
@@ -66,7 +66,6 @@ PACKAGES = ['vip_hci',
             'vip_hci.config',
             'vip_hci.fits',
             'vip_hci.fm',
-            'vip_hci.greedy',
             'vip_hci.invprob',
             'vip_hci.metrics',
             'vip_hci.preproc',
