+++ vip_hci/greedy/pca_local.py
@@ -1,2041 +1,0 @@
-#! /usr/bin/env python
-
-"""
-Module with local/smart PCA (annulus or patch-wise in a multi-processing
-fashion) model PSF subtraction for ADI, ADI+SDI (IFS) and ADI+RDI datasets.
-
-.. [PAI18]
-   | Pairet et al. 2018
-   | **Reference-less algorithm for circumstellar disks imaging**
-   | *arXiv e-print:1812.01333*
-   | `https://arxiv.org/abs/1812.01333
-     <https://arxiv.org/abs/1812.01333>`_
-     
-     
-.. [PAI21]
-   | Pairet et al. 2021
-   | **MAYONNAISE: a morphological components analysis pipeline for circumstellar discs and exoplanets imaging in the near-infrared**
-   | *MNRAS, Volume 503, Issue 3, pp. 3724-3742*
-   | `https://arxiv.org/abs/2008.05170
-     <https://arxiv.org/abs/2008.05170>`_
-   
-"""
-
-__author__ = 'Valentin Christiaens'
-__all__ = ['pca_annular_it',
-           'feves',
-           'feves_auto']
-
-from multiprocessing import cpu_count
-import numpy as np
-from scipy.interpolate import InterpolatedUnivariateSpline
-from ..config.utils_conf import pool_map, iterable
-from .pca_fullfr import _interp2d_rad
-from ..preproc import cube_derotate, cube_shift
-from ..preproc.derotation import _define_annuli
-from ..psfsub import nmf_annular, pca_annular
-from ..metrics import throughput, stim_map, mask_sources, inverse_stim_map
-from ..var import (get_annulus_segments, matrix_scaling, mask_circle,
-                   frame_filter_lowpass, frame_center, dist_matrix)
-# DELETE
-from ..fits import write_fits
-
-def pca_annular_it(cube, angle_list, cube_ref=None, ncomp=1, n_it=10, thr=1., 
-                   thr_per_ann=True, thru_corr=False, n_neigh=0, strategy='ADI', 
-                   psfn=None, n_br=6, radius_int=0, fwhm=4, asize=4, 
-                   n_segments=1, delta_rot=(0.1, 1), svd_mode='lapack', nproc=1, 
-                   min_frames_lib=2, max_frames_lib=200, tol=1e-1, scaling=None, 
-                   imlib='vip-fft', interpolation='lanczos4', collapse='median', 
-                   full_output=False, verbose=True, weights=None,
-                   interp_order=2, rtol=1e-2, atol=1, **rot_options):
-    """
-    Iterative version of annular PCA.
-
-    The algorithm finds significant disc or planet signal in the final PCA
-    image, then subtracts it from the input cube (after rotation) just before
-    (and only for) projection onto the principal components. This is repeated
-    n_it times, which progressively reduces geometric biases in the image
-    (e.g. negative side lobes for ADI, radial negative signatures for SDI).
-    This is similar to the algorithm presented in Pairet et al. (2020).
-
-    The same parameters as pca_annular() can be provided. There are two
-    additional parameters related to the iterative algorithm: the number of
-    iterations (n_it) and the threshold (thr) used for the identification of
-    significant signals.
-
-    Parameters
-    ----------
-    cube : numpy ndarray, 3d or 4d
-        Input cube.
-    angle_list : numpy ndarray, 1d
-        Corresponding parallactic angle for each frame.
-    cube_ref : numpy ndarray, 3d, optional
-        Reference library cube. For Reference Star Differential Imaging.
-    ncomp : 'auto', int, tuple, 1d numpy array or tuple, optional
-        How many PCs are used as a lower-dimensional subspace to project the
-        target (sectors of) frames. Depends on the dimensionality of `cube`.
[…]
