### 2022-08-11 8055563
Added the option to control the radial sampling for calculation of the noise level
