### 2022-08-24 f0319b4
Added the option of multiprocessing for cube shifts within fake companion injection routine
