+++ vip_hci/fm/fakecomp.py
@@ -26,14 +26,14 @@ from ..preproc import (cube_crop_frames, frame_shift, frame_crop, cube_shift,
                        frame_rotate)
 from ..var import (frame_center, fit_2dgaussian, fit_2dairydisk, fit_2dmoffat,
                    get_circle, get_annulus_segments, dist_matrix)
-from ..config.utils_conf import print_precision, check_array
+from ..config.utils_conf import print_precision, check_array, pool_map, iterable
 
 
 def cube_inject_companions(array, psf_template, angle_list, flevel, rad_dists,
-                           plsc=None, n_branches=1, theta=0, imlib='vip-fft',
-                           interpolation='lanczos4', transmission=None,
-                           radial_gradient=False, full_output=False,
-                           verbose=False):
+                           plsc=None, n_branches=1, theta=0, nproc=1, 
+                           imlib='vip-fft', interpolation='lanczos4', 
+                           transmission=None, radial_gradient=False, 
+                           full_output=False, verbose=False):
     """ Injects fake companions in branches, at given radial distances.
 
     Parameters
@@ -72,6 +72,9 @@ def cube_inject_companions(array, psf_template, angle_list, flevel, rad_dists,
         Angle in degrees for rotating the position of the first branch that by
         default is located at zero degrees. Theta counts counterclockwise from
         the positive x axis.
+    nproc: int or None, optional
+        Number of CPUs to use for multiprocessing. If None, will be 
+        automatically set to half the number of available CPUs.
     imlib : str, optional
         See the documentation of the ``vip_hci.preproc.frame_shift`` function.
     interpolation : str, optional
@@ -105,24 +108,24 @@ def cube_inject_companions(array, psf_template, angle_list, flevel, rad_dists,
         [full_output & transmission != None] Array with injected psf affected
         by transmission (serves to check radial transmission)
 
-
     """
+
     def _cube_inject_adi(array, psf_template, angle_list, flevel, plsc,
                          rad_dists, n_branches=1, theta=0, imlib='vip-fft',
                          interpolation='lanczos4', transmission=None,
                          radial_gradient=False, verbose=False):
-
+        if np.isscalar(flevel):
+            flevel = np.ones_like(angle_list)*flevel
+            
         if transmission is not None:
             # last radial separation should be beyond the edge of frame
             interp_trans = interp1d(transmission[0], transmission[1])
 
-        positions = []
-        w = int(np.ceil(size_fc/2))
-        if size_fc % 2:  # new convention
-            w -= 1
-        sty = int(ceny) - w
-        stx = int(cenx) - w
+        ceny, cenx = frame_center(array[0])
+        nframes = array.shape[-3]
+        size_fc = psf_template.shape[-1]
 
+        positions = []
         # fake companion cube
         fc_fr = np.zeros([nframes, size_fc, size_fc])
         if psf_template.ndim == 2:
@@ -152,66 +155,37 @@ def cube_inject_companions(array, psf_template, angle_list, flevel, rad_dists,
                             fc_fr_rad[:, i] = interp_trans(d[i])*fc_fr[:, i]
 
                         # check the effect of transmission on a single PSF tmp
-                        psf_trans = frame_rotate(fc_fr_rad[0],
-                                                 -(ang*180/np.pi-angle_list[0]),
-                                                 imlib=imlib_rot,
-                                                 interpolation=interpolation)
+                        # psf_trans = frame_rotate(fc_fr_rad[0],
+                        #                          -(ang*180/np.pi-angle_list[0]),
[…]
