### 2022-08-24 82bbe7b
Better check of type of ncomp in project_subtract PCA function
