### 2022-08-24 106dd96
Changed location of new parameter nproc in cube_inject_companions for retro-compatibility
