### 2022-08-24 c7e411c
implementation of multiprocessing-based fake companion injection
