### 2022-08-24 64d2a09
Improved description for cube_shift
