### 2022-09-13 2ef0b56
Added progress bar to FT-based bad pixel correction
