### 2022-09-13 7f3aff9
Added printed info during FT-based bad pixel correction for verbose=True
