### 2022-09-14 b38b28e
docstring update for normalize_psf
