### 2022-09-20 #4774 @editorialbot
[REVIEW]: VIP: A Python package for high-contrast imaging
**Submitting author:** @VChristiaens (<a href="http://orcid.org/0000-0002-0101-8814">Valentin Christiaens</a>)
**Repository:** https://github.com/vortex-exoplanet/VIP
**Branch with paper.md** (empty if default branch):
**Version:** v1.3.5
