### 2022-09-30 2f7da85
More specific skimage requirement added
