### 2022-09-30 58f9c9e
Updated printed outputs for FFT-based bad pixel correction
