### 2022-10-05 6737ee2
Made check of input delta_rot parameter format more general
