### 2022-10-11 a4c13ec
minor bug fix when theta_init is required for 4D input cube in no mSDI mode (i.e. ADI in all spectral channels)
