### 2022-10-17 d2a8a98
(frame_average_radprofile) frame with the radial profile subtracted can be returned as second output
