### 2022-10-21 090ab5b
consistent handling of nproc
