### 2022-10-24 #4774 @ivastar
To get started paste the following command as a separate comment: `@editorialbot generate my checklist`
