### 2022-11-14 ff78040
Modified detection function behaviour to match description
