### 2022-11-14 c5640d3
New option in bad pixel functions: correct_only, set by default to False such that a first map of bad pixels (e.g. static ones) can be provided before identifying additional bad pixels
