+++ vip_hci/preproc/badpixremoval.py
@@ -29,7 +29,8 @@ from scipy.ndimage import median_filter
 from astropy.stats import sigma_clipped_stats
 from multiprocessing import cpu_count
 from ..stats import clip_array, sigma_filter
-from ..var import frame_center, get_annulus_segments, frame_filter_lowpass
+from ..var import (frame_center, get_annulus_segments, frame_filter_lowpass, 
+                   frame_filter_highpass)
 from ..config import timing, time_ini, Progressbar
 from ..config.utils_conf import pool_map, iterable
 from .rescaling import find_scal_vector, frame_rescaling
@@ -45,9 +46,10 @@ except ImportError:
     no_numba = True
 
 
-def frame_fix_badpix_isolated(array, bpm_mask=None, sigma_clip=3, num_neig=5,
-                              size=5, protect_mask=0, cxy=None, mad=False,
-                              ignore_nan=True, verbose=True, full_output=False):
+def frame_fix_badpix_isolated(array, bpm_mask=None, correct_only=False, 
+                              sigma_clip=3, num_neig=5, size=5, protect_mask=0, 
+                              cxy=None, mad=False, ignore_nan=True, 
+                              verbose=True, full_output=False):
     """ Corrects the bad pixels, marked in the bad pixel mask. The bad pixel is
     replaced by the median of the adjacent pixels. This function is very fast
     but works only with isolated (sparse) pixels.
@@ -60,6 +62,9 @@ def frame_fix_badpix_isolated(array, bpm_mask=None, sigma_clip=3, num_neig=5,
         Input bad pixel map. Zeros frame where the bad pixels have a value of
         1. If None is provided a bad pixel map will be created using
         sigma clip statistics.
+    correct_only : bool, opt
+        If True and bpix_map is provided, will only correct for provided bad
+        pixels. Else, the algorithm will determine (more) bad pixels.
     sigma_clip : int, optional
         In case no bad pixel mask is provided all the pixels above and below
         sigma_clip*STDDEV will be marked as bad.
@@ -102,6 +107,9 @@ def frame_fix_badpix_isolated(array, bpm_mask=None, sigma_clip=3, num_neig=5,
         raise TypeError('Array is not a 2d array or single frame')
     if size % 2 == 0:
         raise TypeError('Size of the median blur kernel must be an odd integer')
+    if correct_only and bpm_mask is None:
+        msg = "Bad pixel map should be provided if correct_only is True."
+        raise ValueError(msg)
 
     if bpm_mask is not None:
         bpm_mask = bpm_mask.astype('bool')
@@ -120,10 +128,10 @@ def frame_fix_badpix_isolated(array, bpm_mask=None, sigma_clip=3, num_neig=5,
     else:
         cx, cy = cxy
 
-    if bpm_mask is None:
+    if bpm_mask is None or not correct_only:
         ori_nan_mask = np.where(np.isnan(frame))
-        ind = clip_array(frame, sigma_clip, sigma_clip, neighbor=neigh,
-                         num_neighbor=num_neig, mad=mad)
+        ind = clip_array(frame, sigma_clip, sigma_clip, bpm_mask, 
+                         neighbor=neigh, num_neighbor=num_neig, mad=mad)
         bpm_mask = np.zeros_like(frame)
         bpm_mask[ind] = 1
         if ignore_nan:
@@ -149,9 +157,10 @@ def frame_fix_badpix_isolated(array, bpm_mask=None, sigma_clip=3, num_neig=5,
         return array_out
 
 
-def cube_fix_badpix_isolated(array, bpm_mask=None, sigma_clip=3, num_neig=5,
-                             size=5, frame_by_frame=False, protect_mask=0,
-                             cxy=None, mad=False, ignore_nan=True, verbose=True,
[…]
