### 2022-11-14 061da8e
pca_annular changes: scale_list not mandatory for 4D input cube, and new collapse_ifs option added
