### 2022-11-24 #4774 @bethbiller
A few comments for the authors, mostly (but not entirely) on the documentation provided:

1) Installation -- I had no problem installing VIP on two different linux machines, however, I couldn't get it to install on my Mac because I couldn't manage to get opencv-python to install correctly.  The main functionality of the package does not rely on opencv-python and every tool in the package can potentially be run without the opencv-python package.  Thus, it would be really good to enable an installation without opencv-python if possible.

2) There are a number of not-clearly defined acronyms on the documentation pages.  AGPM (Annular Groove Phase Mask) is not defined in the text but is relatively clear by context.  In sections 3.5-3.10 of the documentation, a number of high-contrast imaging algorithms are referred to first by their acronyms -- please spell out the acronym first before using it.

3) Section 4.2.3: Something has gone wrong with the STIM map definition -- it is not displaying correctly:

![Screen Shot 2022-11-24 at 12 03 38 PM](https://user-images.githubusercontent.com/61143641/203780070-d270d0f3-1afd-4883-9a4f-d4ad68d7bb08.png)

4) All the example datasets used in the documentation are provided in a separate VIP_extras repository.  I had a hard time finding this repository, as it is not in the main VIP repository at:

https://github.com/vortex-exoplanet/VIP

but rather is in a separate repository at:

https://github.com/vortex-exoplanet/VIP_extras

A direct link to this repository in the documentation (as opposed to just mentioning the VIP_extras repository) would be very helpful.
