### 2022-11-24 #4774 @ivastar
@m-samland also sent me the following via e-mail:

There is one crucial citation missing in the paper since it mentions that this is the only package that does FFT shifting, rotations etc. There is another high-contrast imaging pipeline that did this before that should at least be mentioned, even if the code is not on github as far as I know.

https://ui.adsabs.harvard.edu/abs/2016MNRAS.455.2178H/abstract
