### 2022-11-25 ae1a033
Additional options for bad pixel correction
