### 2022-11-25 80fbf17
filter kernels now can be provided as tuple of 2 values for different kernel sizes along x and y
