### 2022-11-26 21cf8cc
added missing delta_rot in docstring for mcmc_negfc_sampling
