### 2022-11-26 49dac1c
Updated README for clearer instructions on how to run the jupyter notebook tutorials
