### 2022-11-26 3b48ca8
Modified requirements list (developer dependencies) for continuous integration tests
