### 2022-11-26 12598bf
Bold faced instructions to run tutorial notebooks
