### 2022-11-26 c1429d6
Updated info regarding the opencv optional dependency
