### 2022-11-26 e1d7c0a
Updated paper draft as per referee feedback
