### 2022-11-26 dfdb5b3
Updated paper draft as per referee feedback
