+++ paper.md
@@ -155,7 +155,7 @@ high-contrast imaging algorithms
 Given the rapid expansion of ``VIP``, we summarize here all novelties that were 
 brought to the package over the past five years. Specifically, the rest of this 
 manuscript summarizes all major changes since v0.7.0 [@Gomez:2017], that are 
-included in the latest release of ``VIP`` (v1.3.3). At a structural level, 
+included in the latest release of ``VIP`` (v1.3.4). At a structural level, 
 ``VIP`` underwent a major change since version v1.1.0, which aimed to migrate 
 towards a more streamlined and easy-to-use architecture. The package now 
 revolves around five major modules (`fm`, `invprob`, `metrics`, `preproc` and 
