### 2022-11-28 #4774 @VChristiaens
Thank you @bethbiller and @m-samland for taking the time to review our package! All the changes you suggested are now implemented:

1. The opencv dependency is now set as a developer requirement (requirements-dev.txt instead of requirements.txt), such that a standard VIP installation (e.g. through pip) would not install it. That being said, I have managed to have both opencv and all other VIP dependencies successfully installed on my environment on MacOS (which is why I had switched it to be a regular requirement for convenience). If you wish so, feel free to raise an issue or a question on the [Q/A](https://github.com/vortex-exoplanet/VIP/discussions/categories/q-a) with the details of the problem you are experiencing @bethbiller, so that we can help. Although not mandatory, opencv can still be useful to obtain quick results (typically over an order of magnitude faster for ADI-based algorithms).
2. & 3. Thanks for noting these. All acronyms are now expanded upon their first use between Secs. 3.5 and 3.10. I also skimmed the other notebooks to do the same where relevant. The typo near the STIM map definition is now also corrected.
4. The relevant sentences in both the README and the online documentation providing information on the location of the tutorials are now set in bold font for clarity. Note that the reason why a different GitHub repository is needed for tutorials and test datasets is to avoid a prohibitive importation time for the VIP package (typically over a minute long on my laptop when I tested their inclusion to the main repository).

I have now rephrased the sentence mentioning FFT-based image operations in the paper, to also include a citation to Hagelberg et al. (2016). Thanks for pointing out that work.
