### 2022-11-29 766299d
Added option to provide an azimuthal range to be avoided for noise apertures to be used for the calculation of the SNR
