### 2022-11-30 54df56a
Bug fix for exclude_theta_range option of snr() function
