### 2022-12-06 87e47be
Added contributors to the author list
