### 2022-12-14 325a3ef
cube_fix_badpix_isolated function updated with multiprocessing
