### 2022-12-22 b8d737b
detection returns snr when full_output=True as in docstring
