### 2022-12-23 #4774 @ivastar
@VChristiaens the review is complete! 🎉

Moving towards publishing the submission.

First a couple of minor edits to the text:
- Line 42: "annuli, and considering a parallactic angle" --> "annuli, and to consider a parallactic angle"
- Line 45: "speckle suppression, hence help with the detection" --> speckle suppression, helping with the detection
- Line 75: "novelties" --> "new features"
- Line 78-79: "which aimed to migrate towards a more" --> "which migrated it towards a more"

Second, please make a tagged release and archive, and report the version number and archive DOI in this thread.
When creating the archive, please make sure that the title and author lists match those of the paper.

Let me know if you have questions.
