### 2022-12-26 a1d95f6
Added function to define the independent apertures
