### 2022-12-26 e4110fe
Renamed obj_tmp variables for consistency with routines in other modules
