+++ vip_hci/preproc/badpixremoval.py
@@ -26,9 +26,7 @@ __all__ = ['frame_fix_badpix_isolated',
 import numpy as np
 from skimage.draw import disk, ellipse
 from scipy.ndimage import median_filter
-from astropy.convolution import (convolve, Gaussian2DKernel)
-from astropy.convolution import interpolate_replace_nans as interp_nan
-from astropy.stats import sigma_clipped_stats, gaussian_fwhm_to_sigma
+from astropy.stats import sigma_clipped_stats
 from multiprocessing import cpu_count
 from ..stats import clip_array, sigma_filter
 from ..var import frame_center, get_annulus_segments, frame_filter_lowpass
@@ -210,8 +208,7 @@ def cube_fix_badpix_isolated(array, bpm_mask=None, correct_only=False,
     verbose : bool, optional
         If True additional information will be printed.
     full_output: bool, {False,True}, optional
-        Whether to return as well the cube of bad pixel maps and the cube of
-        defined annuli.
+        Whether to return as well the cube of bad pixel maps.
 
     Return
     ------
@@ -385,7 +382,7 @@ def cube_fix_badpix_annuli(array, fwhm, cy=None, cx=None, sig=5.,
 
     Returns
     -------
-    obj_tmp: 2d or 3d array
+    array_corr: 2d or 3d array
         The bad pixel corrected frame/cube.
     bpix_map: 2d or 3d array
         [full_output=True] The bad pixel map or the cube of bpix maps
@@ -393,21 +390,20 @@ def cube_fix_badpix_annuli(array, fwhm, cy=None, cx=None, sig=5.,
         [full_output=True] The cube of defined annuli
     """
 
-    obj_tmp = array.copy()
-    ndims = obj_tmp.ndim
+    ndims = array.ndim
     assert ndims == 2 or ndims == 3, "Object is not two or three dimensional.\n"
 
     # thresholds
     if min_thr is None:
-        min_thr = np.amin(obj_tmp)-1
+        min_thr = np.amin(array_corr)-1
     if max_thr is None:
-        max_thr = np.amax(obj_tmp)-1
+        max_thr = np.amax(array_corr)-1
 
-    def bp_removal_2d(obj_tmp, cy, cx, fwhm, sig, protect_mask, r_in_std,
+    def bp_removal_2d(array, cy, cx, fwhm, sig, protect_mask, r_in_std,
                       r_out_std, verbose):
 
-        n_x = obj_tmp.shape[1]
-        n_y = obj_tmp.shape[0]
+        n_x = array.shape[1]
+        n_y = array.shape[0]
 
         # Squash the frame if twice less resolved vertically than horizontally
         if half_res_y:
@@ -417,10 +413,10 @@ def cube_fix_badpix_annuli(array, fwhm, cy=None, cx=None, sig=5.,
                 raise ValueError(msg+msg2)
             n_y = int(n_y/2)
             cy = int(cy/2)
-            frame = obj_tmp.copy()
-            obj_tmp = np.zeros([n_y, n_x])
+            frame = array.copy()
+            array = np.zeros([n_y, n_x])
             for yy in range(n_y):
-                obj_tmp[yy] = frame[2*yy]
+                array[yy] = frame[2*yy]
 
         # 1/ Stddev of background
         if r_in_std or r_out_std:
@@ -430,13 +426,13 @@ def cube_fix_badpix_annuli(array, fwhm, cy=None, cx=None, sig=5.,
             else:
                 r_out_std = min(n_y-(cy+r_in_std), cy-r_in_std,
                                 n_x-(cx+r_in_std), cx-r_in_std)
-            ceny, cenx = frame_center(obj_tmp)
+            ceny, cenx = frame_center(array)
             width = max(2, r_out_std-r_in_std)
[…]
