### 2022-12-26 d68875f
Renamed variables for consistency
