### 2022-12-27 93c19bc
Bug fix with propagated argument in frame_fix_badpix_fft
