### 2022-12-27 73de1fd
Added check for padding factor in frame_pad function
