+++ vip_hci/preproc/cosmetics.py
@@ -172,10 +172,10 @@ def frame_pad(array, fac, fillwith=0, loc=0, scale=1, keep_parity=True,
     if not array.ndim == 2:
         raise TypeError("The input array must be 2d")
     if np.isscalar(fac):
-        if fac <1:
+        if fac < 1:
             raise ValueError("fac should be larger than 1")
-    else
-        if fac[0]<1 or fac[-1]<1:
+    else:
+        if fac[0] < 1 or fac[-1] < 1:
             raise ValueError("fac elements should be larger than 1")
 
     y, x = array.shape
