### 2023-01-02 #4774 @VChristiaens
Dear @ivastar,
Thanks for finalizing the review!
I have edited the text as per your suggestions, made a new tagged release for v1.3.5 (the version number was also updated in the manuscript), and archived it on Zenodo. Here is the DOI: 10.5281/zenodo.7499314
Best wishes,
Valentin
