### 2023-01-03 79eba2e
Bug fix for bad pixel correction by interpolation in 2D input frame
