### 2023-01-03 38dc8b6
Small edits to JOSS paper
