### 2023-01-04 350a1c9
Dummy call added on cube_fix_badpix_isolated and cube_correct_nan functions
