### 2023-01-04 #4774 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
